% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-aichi-districts.R
\docType{data}
\name{aichi_districts}
\alias{aichi_districts}
\title{Aichi prefecture administrative district data}
\format{
Each is a tibble with 69 rows 6 variables:
\describe{
\item{city}{City names}
\item{city_code}{City codes}
\item{city_kanji}{City names in japanese}
\item{geom}{Geometry data (polygon or multi-polygon type) of boundaries}
}
}
\usage{
aichi_districts
}
\description{
Information on administrative areas such as cities and wards
in Aichi prefecture.
}
\examples{
aichi_districts
}
\keyword{datasets}
