% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aichicities-package.R
\docType{package}
\name{aichicities-package}
\alias{aichicities}
\alias{aichicities-package}
\title{aichicities: Aichi Prefecture administrative district data}
\description{
This package provides administrative district data of Aichi Prefecture, Japan.
}
\author{
\strong{Maintainer}: Keisuke ANDO \email{ando@maslab.aitech.ac.jp}

}
\keyword{internal}
