% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-aichi-district-data.R
\name{download_aichi_district_data}
\alias{download_aichi_district_data}
\title{Download the Aichi Prefecture administrative district data}
\usage{
download_aichi_district_data(download_dir = getwd())
}
\arguments{
\item{download_dir}{A directory where downloaded file is to saved.}
}
\value{
Path of the downloaded file (invisibly).
}
\description{
Download the Aichi Prefecture administrative district data
provided by the Ministry of Land, Infrastructure, Transport and Tourism.
}
\examples{
\dontrun{
download_aichi_district_data("download-dir-path")
}
}
