% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-district.R, R/find_district.R
\name{find_districts}
\alias{find_districts}
\title{Find district by city name}
\usage{
find_districts(city_name)

find_districts(city_name)
}
\arguments{
\item{city_name}{Part of the city name}
}
\value{
districts of the specified city name

districts of the specified city name
}
\description{
The function returns the districts of the specified city name.
The search is performed with partial matches.

The function returns the districts of the specified city name.
The search is performed with partial matches.
}
\examples{
\dontrun{
toyota_districts <- find_districts("Toyota-shi")
}
\dontrun{
toyota_districts <- find_districts("Toyota-shi")
}
}
