% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-aichi-district-data.R
\name{read_aichi_district_data}
\alias{read_aichi_district_data}
\title{Read the Aichi Prefecture administrative district data}
\format{
Each is a tibble with 69 rows 6 variables:
\describe{
\item{city}{City names}
\item{city_code}{City codes}
\item{city_kanji}{City names in japanese}
\item{geom}{Geometry data (polygon or multi-polygon type) of boundaries}
}
}
\usage{
read_aichi_district_data(file)
}
\arguments{
\item{file}{A file path of data to read.}
}
\value{
Information on administrative areassuch as cities and wards
in Aichi prefecture.
}
\description{
Read the Aichi Prefecture administrative district data.
}
\examples{
\dontrun{
aichi_districts <- read_aichi_district_data("N03-20210101_23_GML.zip")
}
}
